refactor: split stmt.rs (427 lines) into stmt/ directory module
stmt.rs was 27 lines over the 400-line project limit and would grow
further with Phase 6.1 try-catch. Split into focused sub-modules:

- stmt/mod.rs (116 lines): dispatcher + simple arms (return, expr, block, if, throw)
- stmt/var_decl.rs (148 lines): const/let with ident, object/array destructuring
- stmt/loops.rs (126 lines): while, for-of, for-in, for, do-while
- stmt/switch.rs (47 lines): switch → match conversion

Extracted shared helpers: extract_loop_var() and ensure_block() reduce
duplication across loop converters. Variable declaration logic split into
convert_ident_decl, convert_object_destructuring, convert_array_destructuring.

No behavioral changes. All 158 tests pass.

Closes #30
